requirements.txt
django==5.1.7
djangorestframework==3.15.2
environ==1.0
psycopg2==2.9.10
requests
celery==5.4.0
redis==5.2.1
python-dotenv